+++ examples/__init__.py
@@ -21,4 +21,4 @@ __author__ = "Ian Housman"
 __email__ = "ian.housman@gmail.com"
 
 # Version format yyyy.m.n
-__version__ = "2024.11.2"
+__version__ = "2024.11.3"